% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{example.trial.data}
\alias{example.trial.data}
\title{A function to create example trial-level data}
\usage{
example.trial.data(
  seed = NULL,
  mean.RT = 400,
  sd.RT = 150,
  mean.accuracy = 0.9,
  condition = c(rep("A", 5), rep("B", 5)),
  shuffle = T,
  ID = 1:8,
  session = 1:2
)
}
\arguments{
\item{mean.RT, sd.RT}{Scalar, overall mean and standard deviation values being passed to rnorm() to create example reactiont times. Any negative values generated by rnorm() will be made positive. Default to 400 and 150, respectively.}

\item{mean.accuracy}{Scalar, overall mean accuracy used to create example accuracies. Defaults to 0.9.}

\item{condition, shuffle}{A vector of trial conditions, shuffled (T) or not shuffled (F) for each participant and each session. Is being used to determine the number of trials for each participant, repeated for each session. Defaults to c(rep('A',5), rep('B',5)).}

\item{ID}{A vector of participant identifiers. Defaults to '1:10'}

\item{session}{A vector of session identifiers. Defaults to '1:2'}
}
\description{
A function to create example trial-level local-global data from some default experiment to be used for example calculations
}
\examples{
example.trial.data()
}
\keyword{data}
\keyword{example,}
