% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{fill.missing.values}
\alias{fill.missing.values}
\title{Fill up missing values in an ldf}
\usage{
fill.missing.values(ldf, fill.by = NULL, notify = T, value.col = "value")
}
\arguments{
\item{ldf}{A long format dataframe with one column holding the values and at least one column holding identifiers that can be used to group the values. Compare example.trial.data().}

\item{fill.by}{Identifier-columns that are used to create groupings, which are filled up. Defaults to NULL, which means all columns besides 'value.col' are used.}

\item{notify}{Provides progress information. Defaults to TRUE.}

\item{value.col}{Name of the column holding the values in the ldf. Defaults to 'value'.}
}
\description{
Fill missing rows to make sure every combination of the grouping-identifiers, measure-identifiers, and observations, exists (measure and observation can be disabled)
}
\examples{
# Introduce some missing rows into the example data
ldf = example.trial.data(); ldf = ldf[sort(sample(1:nrow(ldf), 0.6*nrow(ldf))),]
# Fill the rows of all missing groupings
ldf2 = fill.missing.values(ldf = ldf)
# Note that not specifying 'fill.by', as done before, leads to non-desired output in the example dataset, because each value is already fully specified with the columns session,ID,trial,measure. The column condition introduces over-specification, since when trial 1 for participant 1 in session 1 has EITHER condition A OR condition B. If this row 1,1,1 is missing, it needs to be added but with condition NA, because the condition of the trial is not known. Therefore, condition should not be included in 'fill.by'.
ldf2 = fill.missing.values(ldf = ldf, fill.by = c('ID','session','trial','measure'))
}
\keyword{data}
