% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trim.by_spread}
\alias{trim.by_spread}
\title{Trim values beyond a defined spread in groupings of an ldf}
\usage{
trim.by_spread(
  ldf,
  group.by,
  apply.to = NULL,
  apply.not.to = NULL,
  center.func = "mean",
  spread.func = "sd",
  trim = 2.5,
  report.by = NULL,
  return.report = F,
  return.stats = F,
  notify = T,
  value.col = "value"
)
}
\arguments{
\item{ldf}{A long format dataframe with a column (default: 'value') holding the values, a column (default: 'measure', e.g. reaction time) describing the value, a column (default: 'observation') linking different measures of the same observation together, and at least one further column being used as identifier-column (such as 'participant ID'). Use example.trial.data() for a correctly formatted long dataframe. Tip: you can use dplyr::pivot_longer to transform a wide dataframe into a long dataframe.}

\item{group.by}{Identifier-columns that are used to create groupings, to which the test is applied.}

\item{apply.to, apply.not.to}{A named list of vectors to (un)select rows. The names of the columns are specified by the element names, and the identifiers in this column to which the trimming should (not) be applied are specified by the vector. Default to NULL, which means no (un)selection is applied to any column.}

\item{center.func, spread.func}{Name of a function that can take a vector and output a scalar. These function are used to summarise the values of each grouping and find their 'central point' and spread, respectively. Defaults to 'mean' and 'sd', respectively. 'na.rm = T' is passed as additional argument.}

\item{trim}{Vector of one or two values. Multiplied with the spread and subtracted/added to the central tendency to get the lower and upper bound. A single value is duplicated. Defaults to c(2.5,2.5).}

\item{report.by}{Identifier-columns that are used to create groupings, for which statistics are printed when notify is TRUE. Defaults to NULL, which means one value for the entire ldf is returned.}

\item{return.report, return.stats}{Adds the trimming report/statistics of centrality and spread to the trimmed ldf in a list. Defaults to FALSE.}

\item{notify}{Provides progress information. Defaults to TRUE.}

\item{value.col}{Name of the column holding the values in the ldf. Defaults to 'value'.}
}
\description{
Groups an ldf by certain columns, calculates a freely definable centrality and spread statistic for each grouping, and trims values that are more than x times smaller or larger than the spread by overwriting them with NA.
}
\examples{
# Trimming each participants RTs that are more than 2 SD beyond the mean of this participant
ldf = trim.by_spread(ldf = example.trial.data(), group.by = 'ID', apply.to = list(measure = 'RT'), trim = 2)
# Trimming each participants RTs that are beyond 2 SD, but only above, not below (using Inf in 'trim')
ldf = trim.by_spread(ldf = example.trial.data(), group.by = 'ID', apply.to = list(measure = 'RT'), trim = c(Inf,2))
# Do the same, but now for each session separately and report trimming statistics it in more detail
ldf = trim.by_spread(ldf = example.trial.data(), group.by = c('ID','session'), apply.to = list(measure = 'RT'), trim = c(Inf,2), report.by = c('session','condition'))
}
\keyword{trimming}
