% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trim.statistics}
\alias{trim.statistics}
\title{Return statistics about NA-values per grouping in an ldf}
\usage{
trim.statistics(
  ldf,
  ldf0 = NULL,
  report.by,
  apply.to = NULL,
  apply.not.to = NULL,
  print.stats = F,
  value.col = "value"
)
}
\arguments{
\item{ldf}{A long format dataframe with one column holding the values and at least one column holding identifiers that can be used to group the values. Compare example.trial.data().}

\item{ldf0}{A previous state of ldf, to calculate the difference in NA-values}

\item{report.by}{Identifier-columns that are used to create groupings, for which statistics are returned.}

\item{apply.to, apply.not.to}{A named list of vectors to (un)select rows. The names of the columns are specified by the element names, and the identifiers in this column to which the trimming should (not) be applied are specified by the vector. Default to NULL, which means no (un)selection is applied to any column.}

\item{print.stats}{Prints the stats instead of returning them. Defaults to FALSE.}

\item{value.col}{Name of the column holding the values in ldf. Defaults to 'value'.}
}
\description{
Calculate statistics about NA-values in a specified groupings of a long format dataframe, as well as optionally the difference to a prior state of the ldf (ldf0).
}
\keyword{keyword}
